% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{autocor}
\alias{autocor}
\title{Autocorrelation matrix}
\usage{
autocor(p, rho)
}
\arguments{
\item{p}{dimension of matrix}

\item{rho}{correlation}
}
\description{
Creates autocorrelation matrix of size p with parameter rho
}
\seealso{
\code{\link{GenerateData}}
}
