% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{blockcor}
\alias{blockcor}
\title{Block diagonal matrix}
\usage{
blockcor(rho, groupind)
}
\arguments{
\item{rho}{correlation}

\item{groupind}{group index which indicates belonging to which block.}
}
\description{
Creates block matrix (p by p) using group index (of length p) and different parameter rho for each group.
}
\seealso{
\code{\link{GenerateData}}
}
