% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{bridgeF}
\alias{bridgeF}
\alias{bridge_select}
\alias{bridgeF_bc}
\alias{bridgeF_cb}
\alias{bridgeF_tb}
\alias{bridgeF_bt}
\alias{bridgeF_tc}
\alias{bridgeF_ct}
\alias{bridgeF_bb}
\alias{bridgeF_tt}
\title{Bridge functions}
\usage{
bridge_select(type1 = "trunc", type2 = "continuous")

bridgeF_bc(r, zratio1, zratio2 = NULL)

bridgeF_cb(r, zratio1 = NULL, zratio2)

bridgeF_tb(r, zratio1, zratio2)

bridgeF_bt(r, zratio1, zratio2)

bridgeF_tc(r, zratio1, zratio2 = NULL)

bridgeF_ct(r, zratio1 = NULL, zratio2)

bridgeF_bb(r, zratio1, zratio2)

bridgeF_tt(r, zratio1, zratio2)
}
\arguments{
\item{type1}{Type of the dataset \eqn{X_{1}}. Choose among "continous", "binary", "trunc".}

\item{type2}{Type of the dataset \eqn{X_{2}}.}

\item{r}{correlation.}

\item{zratio1}{proportion of zero values of dataset \eqn{X_{1}}.}

\item{zratio2}{proportion of zero values of dataset \eqn{X_{2}}.}
}
\description{
Bridge function \eqn{F} connects Kendall's tau \eqn{\tau} estimation to unbiased latent correlation matrix \eqn{\Sigma} under Gaussian copula model. There are three possible types of variables: continuos, binary, truncated continuous. Depending on the combination of different types of variables, different bridge function should be applied. \code{bridge_select} shows which bridge function should be used based on data types. Taking the first letter from the type of the variables, the combination of the variables' type follows the bridgeF. For example, bc indicates the first dataset is binary and the second dataset is continuous.
\deqn{ F( \Sigma_{jk} ) = \tau_{jk} }
The input \eqn{r} is an element of \eqn{\Sigma} and the bridge function is calculated elementwise.
}
\examples{

set.seed(1)
n <- 100 # sample size
r <- 0.8 # true correlation

# Data generation

# X1: truncated continuous, X2: continuous
Z <- mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1, r, r, 1), nrow = 2))
X1 <- Z[,1]
X1[Z[,1] < 1] <- 0
X2 <- Z[,2]
zratio1 <- mean(X1==0)
# Choose bridge function
bridge <- bridge_select(type1 = "trunc", type2 = "continuous")
K12 <- KendallTau(X1, X2)
K12
f1 <- function(r)(bridge(r, zratio1 = zratio1) - K12)^2
optimize(f1, lower = -0.99, upper = 0.99, tol = 0.001)[[1]]

# X1: truncated continuous, X2: truncated continuous
X2[Z[,2] < 1] <- 0
zratio2 <- mean(X2==0)
bridge <- bridge_select(type1 = "trunc", type2 = "trunc")
K12 <- KendallTau(X1, X2)
K12
f1 <- function(r)(bridge(r, zratio1 = zratio1, zratio2 = zratio2) - K12)^2
optimize(f1, lower = -0.99, upper = 0.99, tol = 0.001)[[1]]

# X1: truncated continuous, X2: binary
X2[X2 != 0] <- 1
zratio2 <- mean(X2==0)
bridge <- bridge_select(type1 = "trunc", type2 = "binary")
K12 <- KendallTau(X1, X2)
K12
f1 <- function(r)(bridge(r, zratio1 = zratio1, zratio2 = zratio2) - K12)^2
optimize(f1, lower = -0.99, upper = 0.99, tol = 0.001)[[1]]
}
