% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{estimateR}
\alias{estimateR}
\title{Estimate latent correlation matrix based on Kendall's tau}
\usage{
estimateR(X, type = "trunc", rho = 0.01)
}
\arguments{
\item{X}{numeric data matrix (n by p)}

\item{type}{the type of data matrix \code{X} among "continuous", "binary", "trunc".}

\item{rho}{shrinkage level to make correlation matrix estimator positive definite. The default is 0.01.}
}
\value{
A data.frame containing
\itemize{
      \item{type: }{type of the data matrix X}
      \item{R: }{Estimated latent correlation matrix}
}
}
\description{
Estimate latent correlation matrix based on Kendall's tau
}
