% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{estimateR_mixed}
\alias{estimateR_mixed}
\title{Estimate latent correlation matrix based on Kendall's tau for mixed type data}
\usage{
estimateR_mixed(X1, X2, type1 = "trunc", type2 = "continuous", rho = 0.01)
}
\arguments{
\item{X1}{numeric data matrix (n by p1).}

\item{X2}{numeric data matrix (n by p2).}

\item{type1}{the type of data matrix \code{X1} among "continuous", "binary", "trunc".}

\item{type2}{the type of data matrix \code{X2} among "continuous", "binary", "trunc".}

\item{rho}{shrinkage level to make correlation matrix estimator positive definite. The default is 0.01.}
}
\value{
A data.frame containing
\itemize{
      \item{type1: }{type of the data matrix X1}
      \item{type2: }{type of the data matrix X2}
      \item{R: }{Estimated latent correlation matrix of X = cbind(X1, X2)}
      \item{R1: }{Estimated latent correlation matrix of X1}
      \item{R2: }{Estimated latent correlation matrix of X2}
      \item{R12: }{Estimated latent correlation matrix of X12}
}
}
\description{
Estimate latent correlation matrix based on Kendall's tau for mixed type data
}
