% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallCCA.R
\name{find_w12bic}
\alias{find_w12bic}
\title{Internal mixedCCA function

wrapper function for lassobic function in cpp}
\usage{
find_w12bic(n, R1, R2, R12, lamseq1, lamseq2, w1init, w2init, BICtype,
  maxiter = 100, tol = 0.001, verbose = FALSE, convcheck = TRUE,
  addstep = TRUE)
}
\arguments{
\item{n}{Sample size}

\item{R1}{Correlation matrix of dataset \code{X1} (p1 by p1)}

\item{R2}{Correlation matrix of dataset \code{X2} (p2 by p2)}

\item{R12}{Correlation matrix between the dataset \code{X1} and the dataset \code{X2} (p1 by p2)}

\item{lamseq1}{a sequence of lambda values for the datasets \code{X1}. It can be a scalar (a vector of one value). should be the same length with lamseq2.}

\item{lamseq2}{a sequence of lambda values for the datasets \code{X2}. It can be a scalar (a vector of one value). should be the same length with lamseq1.}

\item{w1init}{An initial vector of length p1 for canonical direction \eqn{w1}.}

\item{w2init}{An initial vector of length p1 for canonical direction \eqn{w2}.}

\item{BICtype}{Either 1 or 2: For more details for two options, see the reference.}

\item{maxiter}{The maximum number of iterations allowed.}

\item{tol}{The desired accuracy (convergence tolerance).}

\item{verbose}{If \code{verbose = TRUE}, error values in each iteration will be printed. The default value is \code{FALSE}.}

\item{convcheck}{If \code{convcheck = TRUE}, the convergence error will be printed when the convergence is failed after the algorithm reached \code{maxiter}. The default value is \code{TRUE}.}

\item{addstep}{When the convergence is failed, \code{addstep = TRUE} will try further fine lambda sequence values between the oscillation points. The default value is \code{TRUE}.}
}
\description{
Internal mixedCCA function

wrapper function for lassobic function in cpp
}
