% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromKtoR.R
\name{fromKtoR}
\alias{fromKtoR}
\title{From Kendall's tau to correlation matrix estimator using bridge function}
\usage{
fromKtoR(K, zratio = NULL, type = "trunc", tol = 0.001)
}
\arguments{
\item{K}{Kendall's tau correlation matrix (p by p) of \eqn{X}.}

\item{zratio}{proportion of zero values of dataset \eqn{X}.}

\item{type}{type of the dataset \eqn{X}. Choose among "continous", "binary", "trunc".}

\item{tol}{the desired accuracy (convergence tolerance).}
}
\description{
From Kendall's tau to correlation matrix estimator using bridge function
}
\examples{
library(mixedCCA)

set.seed(1)

n <- 100 # sample size
r <- 0.8 # true correlation

# Data generation (X: truncated continuous)
Z <- mvrnorm(n, mu = c(0, 1, 0), Sigma = matrix(c(1, r, r, r, 1, r, r, r, 1), nrow = 3))
X <- ifelse(Z > 0, Z, 0)
zratio <- colMeans(X==0)
K <- Kendall_matrix(X)
fromKtoR(K, zratio = zratio, type = "trunc")

}
