% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromKtoR.R
\name{fromKtoR_mixed}
\alias{fromKtoR_mixed}
\title{From Kendall's tau to correlation matrix estimator using bridge function for mixed type data}
\usage{
fromKtoR_mixed(K12, zratio1 = NULL, zratio2 = NULL, type1 = "trunc",
  type2 = "continuous", tol = 0.001)
}
\arguments{
\item{K12}{Kendall's tau correlation matrix (p1 by p2) between \eqn{X1} and \eqn{X2}.}

\item{zratio1}{proportion of zero values of dataset \eqn{X1}.}

\item{zratio2}{proportion of zero values of dataset \eqn{X2}.}

\item{type1}{type of the dataset \eqn{X1} among "continous", "binary", "trunc".}

\item{type2}{type of the dataset \eqn{X2} among "continous", "binary", "trunc".}

\item{tol}{the desired accuracy (convergence tolerance).}
}
\description{
From Kendall's tau to correlation matrix estimator using bridge function for mixed type data
}
\examples{

library(mixedCCA)
set.seed(1)

n <- 100 # sample size
r <- 0.8 # true correlation

# Data generation (X1: truncated continuous, X2: continuous)
Z <- mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1, r, r, 1), nrow = 2))
X1 <- Z[,1]
X1[Z[,1] < 1] <- 0
X2 <- Z[,2]
zratio1 <- mean(X1==0)
bridge <- bridge_select(type1 = "trunc", type2 = "continuous")
K12 <- KendallTau(X1, X2)
# estimated latent correlation R based on K12.
fromKtoR_mixed(K12, zratio1 = zratio1, type1 = "trunc", type2 = "continuous")

# X1: truncated continuous, X2: truncated continuous
X2[Z[,2] < 1] <- 0
zratio2 <- mean(X2==0)
bridge <- bridge_select(type1 = "trunc", type2 = "trunc")
K12 <- KendallTau(X1, X2)
K12
# estimated latent correlation R based on K12.
fromKtoR_mixed(K12, zratio1 = zratio1, zratio2 = zratio2, type1 = "trunc", type2 = "trunc")

# X1: truncated continuous, X2: binary
X2[X2 != 0] <- 1
zratio2 <- mean(X2==0)
bridge <- bridge_select(type1 = "trunc", type2 = "binary")
K12 <- KendallTau(X1, X2)
K12
# estimated latent correlation R based on K12.
fromKtoR_mixed(K12, zratio1 = zratio1, zratio2 = zratio2, type1 = "trunc", type2 = "binary")

}
