% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallCCA.R
\name{mixedCCA}
\alias{mixedCCA}
\title{Sparse CCA for data of mixed types with BIC criterion}
\usage{
mixedCCA(X1, X2, type1, type2, lamseq1 = NULL, lamseq2 = NULL,
  initlam1 = NULL, initlam2 = NULL, nlambda = 20, lam.eps = 0.01,
  w1init = NULL, w2init = NULL, BICtype, KendallR = NULL, tol = 0.001,
  maxiter = 1000)
}
\arguments{
\item{X1}{A numeric data matrix (n by p1).}

\item{X2}{A numeric data matrix (n by p2).}

\item{type1}{A type of data \code{X1} among "continuous", "binary", "trunc".}

\item{type2}{A type of data \code{X2} among "continuous", "binary", "trunc".}

\item{lamseq1}{A tuning parameter sequence for \code{X1}. The length should be the same as \code{lamseq2}.}

\item{lamseq2}{A tuning parameter sequence for \code{X2}. The length should be the same as \code{lamseq1}.}

\item{initlam1}{An initial value to generate a tuning parameter sequence for \code{X1}, which is the maximum value of tuning parameter grid.}

\item{initlam2}{An initial value to generate a tuning parameter sequence for \code{X2}, which is the maximum value of tuning parameter grid.}

\item{nlambda}{The number of tuning parameter sequence lambda - default is 20.}

\item{lam.eps}{A ratio of the smallest value for lambda to the maximum value of lambda.}

\item{w1init}{An initial vector of length p1 for canonical direction \eqn{w1}.}

\item{w2init}{An initial vector of length p2 for canonical direction \eqn{w2}.}

\item{BICtype}{Either 1 or 2: For more details for two options, see the reference.}

\item{KendallR}{An estimated Kendall \eqn{\tau} matrix. The default is NULL, which means that it will be automatically estimated by Kendall's \eqn{\tau} estimator unless the user supplies.}

\item{tol}{The desired accuracy (convergence tolerance).}

\item{maxiter}{The maximum number of iterations allowed.}
}
\value{
\code{mixedCCA} returns a data.frame containing
\itemize{
      \item{KendallR: }{estimated Kendall's \eqn{\tau} matrix estimator.}
      \item{lambda_seq: }{the values of \code{lambda} used for sparse CCA.}
      \item{w1: }{estimated canonical direction \eqn{w1}.}
      \item{w2: }{estimated canonical direction \eqn{w2}.}
      \item{cancor: }{estimated canonical correlation.}
      \item{selected_x1: }{indices of selected variables in \code{X1}.}
      \item{selected_x2: }{indices of selected variables in \code{X2}.}
}
}
\description{
Applies sparse canonical correlation analysis (CCA) for high-dimensional data of mixed types (continuous/biary/truncated continuous). Derived rank-based estimator instead of sample correlation matrix is implemented. There are two types of BIC criteria for variable selection. We found that BIC1 works best for variable selection, whereas BIC2 works best for prediction.
}
\examples{
# Generate Data
n <- 100; p1 <- 5; p2 <- 7
maxcancor <- 0.9
Sigma1 <- autocor(p1, 0.7)
groupind <- c(rep(1, 2), rep(2, p2-2))
Sigma2 <- blockcor(0.7, groupind)
copula1 <- "exp"; copula2 <- "cube"
mu <- rep(0, p1+p2)
type1 <- type2 <- "trunc"
c1 <- rep(0, p1); c2 <- rep(0, p2)
trueidx1 <- c(0, 0, 1, 1, 1)
trueidx2 <- c(1, 0, 1, 0, 0, 1, 1)
simdata <- GenerateData(n=n, trueidx1 = trueidx1, trueidx2 = trueidx2, maxcancor = maxcancor,
                        Sigma1 = Sigma1, Sigma2 = Sigma2,
                        copula1 = copula1, copula2 = copula2,
                        muZ = mu,
                        type1 = type1, type2 = type2, c1 = c1, c2 = c2
)

X1 <- simdata$X1
X2 <- simdata$X2

# Kendall CCA with BIC1
kendallcca1 <- mixedCCA(X1, X2, type1 = type1, type2 = type2, BICtype = 1)

# Kendall CCA with BIC2. Estimated correlation matrix is plugged in from the above result.
R <- kendallcca1$KendallR
kendallcca2 <- mixedCCA(X1, X2, type1 = type1, type2 = type2, KendallR = R, BICtype = 2)
}
\references{
Yoon G., Carroll R.J. and Gaynanova I. (2018+) \href{http://arxiv.org/abs/1807.05274}{"Sparse semiparametric canonical correlation analysis for data of mixed types"}.
}
