% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallCCA.R
\name{standardCCA}
\alias{standardCCA}
\title{Classical CCA using correlation matrix estimator.}
\usage{
standardCCA(S1, S2, S12, tol = 1e-04)
}
\arguments{
\item{S1}{correlation matrix of \eqn{X1}.}

\item{S2}{correlation matrix of \eqn{X2}.}

\item{S12}{correlation matrix between \eqn{X1} and \eqn{X2}.}

\item{tol}{cutoff for eigenvalues. During the calculation, eigenvalues smaller than cutoff are set to zero.}
}
\value{
A data.frame containing
\itemize{
      \item{cancor: }{estimated canonical correlation.}
      \item{w1: }{estimated canonical direction \eqn{w1}.}
      \item{w2: }{estimated canonical direction \eqn{w2}.}
}
}
\description{
Classical CCA using correlation matrix estimator.
}
