% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{bridge_select}
\alias{bridge_select}
\alias{bridge}
\alias{Bridge}
\alias{function}
\alias{\eqn{F}}
\alias{connects}
\alias{Kendall's}
\alias{tau}
\alias{\eqn{\tau}}
\alias{estimation}
\alias{to}
\alias{unbiased}
\alias{latent}
\alias{correlation}
\alias{matrix}
\alias{\eqn{\Sigma}}
\alias{under}
\alias{Gaussian}
\alias{copula}
\alias{model.}
\alias{There}
\alias{are}
\alias{three}
\alias{possible}
\alias{types}
\alias{of}
\alias{variables:}
\alias{continuos,}
\alias{binary,}
\alias{truncated}
\alias{continuous.}
\alias{Depending}
\alias{on}
\alias{the}
\alias{combination}
\alias{different}
\alias{variables,}
\alias{should}
\alias{be}
\alias{applied.}
\alias{\code{bridge_select}}
\alias{shows}
\alias{which}
\alias{used}
\alias{based}
\alias{data}
\alias{types.}
\alias{Taking}
\alias{first}
\alias{letter}
\alias{from}
\alias{type}
\alias{variables'}
\alias{follows}
\alias{bridgeF.}
\alias{For}
\alias{example,}
\alias{bc}
\alias{indicates}
\alias{dataset}
\alias{is}
\alias{binary}
\alias{and}
\alias{second}
\alias{\deqn{}
\alias{F(}
\alias{\Sigma_{jk}}
\alias{)}
\alias{=}
\alias{\tau_{jk}}
\alias{}}
\alias{The}
\alias{input}
\alias{\eqn{r}}
\alias{an}
\alias{element}
\alias{calculated}
\alias{elementwise.}
\alias{bridgeF_bc}
\alias{bridgeF_cb}
\alias{bridgeF_tb}
\alias{bridgeF_bt}
\alias{bridgeF_tc}
\alias{bridgeF_ct}
\alias{bridgeF_bb}
\alias{bridgeF_tt}
\title{Bridge functions}
\usage{
bridge_select(type1 = "trunc", type2 = "continuous")

bridgeF_bc(r, zratio1, zratio2 = NULL)

bridgeF_cb(r, zratio1 = NULL, zratio2)

bridgeF_tb(r, zratio1, zratio2)

bridgeF_bt(r, zratio1, zratio2)

bridgeF_tc(r, zratio1, zratio2 = NULL)

bridgeF_ct(r, zratio1 = NULL, zratio2)

bridgeF_bb(r, zratio1, zratio2)

bridgeF_tt(r, zratio1, zratio2)
}
\arguments{
\item{type1}{Type of the dataset \eqn{X_{1}}. Choose among "continous", "binary", "trunc".}

\item{type2}{Type of the dataset \eqn{X_{2}}.}

\item{r}{correlation.}

\item{zratio1}{proportion of zero values of dataset \eqn{X_{1}}.}

\item{zratio2}{proportion of zero values of dataset \eqn{X_{2}}.}
}
\description{
Bridge functions
}
\examples{

set.seed(1)
n <- 100 # sample size
r <- 0.8 # true correlation

# Data generation

# X1: truncated continuous, X2: continuous
Z <- mvrnorm(n, mu = c(0, 0), Sigma = matrix(c(1, r, r, 1), nrow = 2))
X1 <- Z[,1]
X1[Z[,1] < 1] <- 0
X2 <- Z[,2]
zratio1 <- mean(X1==0)
# Choose bridge function
bridge <- bridge_select(type1 = "trunc", type2 = "continuous")
K12 <- KendallTau(X1, X2)
K12
f1 <- function(r)(bridge(r, zratio1 = zratio1) - K12)^2
optimize(f1, lower = -0.99, upper = 0.99, tol = 0.001)[[1]]

# X1: truncated continuous, X2: truncated continuous
X2[Z[,2] < 1] <- 0
zratio2 <- mean(X2==0)
bridge <- bridge_select(type1 = "trunc", type2 = "trunc")
K12 <- KendallTau(X1, X2)
K12
f1 <- function(r)(bridge(r, zratio1 = zratio1, zratio2 = zratio2) - K12)^2
optimize(f1, lower = -0.99, upper = 0.99, tol = 0.001)[[1]]

# X1: truncated continuous, X2: binary
X2[X2 != 0] <- 1
zratio2 <- mean(X2==0)
bridge <- bridge_select(type1 = "trunc", type2 = "binary")
K12 <- KendallTau(X1, X2)
K12
f1 <- function(r)(bridge(r, zratio1 = zratio1, zratio2 = zratio2) - K12)^2
optimize(f1, lower = -0.99, upper = 0.99, tol = 0.001)[[1]]
}
