% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateData.R
\name{corrmat}
\alias{corrmat}
\alias{autocor}
\alias{blockcor}
\title{Construct a correlation matrix}
\usage{
autocor(p, rho)

blockcor(rho, groupind)
}
\arguments{
\item{p}{Dimension of matrix}

\item{rho}{Correlation}

\item{groupind}{Group index indicating belonging to which block.}
}
\description{
Functions to create autocorrelation matrix of size p with parameter rho and block correlation matrix (p by p) using group index (of length p) and different parameter rho for each group.
}
\seealso{
See \code{\link{GenerateData}} for an example.
}
