% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateR.R
\name{estimateR}
\alias{estimateR}
\alias{estimateR_mixed}
\title{Estimate latent correlation matrix}
\usage{
estimateR(X, type = "trunc", use.nearPD = TRUE, rho = 0.01,
  tol = 0.001)

estimateR_mixed(X1, X2, type1 = "trunc", type2 = "continuous",
  use.nearPD = TRUE, rho = 0.01, tol = 0.001)
}
\arguments{
\item{X}{A numeric data matrix (n by p), n is the sample size and p is the number of variables.}

\item{type}{A type of variables in \code{X}, must be one of "continuous", "binary" or "trunc".}

\item{use.nearPD}{A logical value indicating whether to use \link[Matrix]{nearPD} or not when the resulting correlation estimator is not positive definite (have at least one negative eigenvalue).}

\item{rho}{Shrinkage parameter for correlation matrix, must be between 0 and 1, the default value is 0.01.}

\item{tol}{Desired accuracy when calculating the solution of bridge function.}

\item{X1}{A numeric data matrix (n by p1).}

\item{X2}{A numeric data matrix (n by p2).}

\item{type1}{A type of variables in \code{X1}, must be one of "continuous", "binary" or "trunc".}

\item{type2}{A type of variables in \code{X2}, must be one of "continuous", "binary" or "trunc".}
}
\value{
\code{estimateR} returns
\itemize{
      \item{type: }{Type of the data matrix \code{X}}
      \item{R: }{Estimated p by p latent correlation matrix of \code{X}}
}

\code{estimateR_mixed} returns
\itemize{
      \item{type1: }{Type of the data matrix \code{X1}}
      \item{type2: }{Type of the data matrix \code{X2}}
      \item{R: }{Estimated latent correlation matrix of whole \code{X} = (\code{X1}, \code{X2}) (p1+p2 by p1+p2)}
      \item{R1: }{Estimated latent correlation matrix of \code{X1} (p1 by p1)}
      \item{R2: }{Estimated latent correlation matrix of \code{X2} (p2 by p2)}
      \item{R12: }{Estimated latent correlation matrix between \code{X1} and \code{X2} (p1 by p2)}
}
}
\description{
Estimation of latent correlation matrix from observed data of (possibly) mixed types (continuous/biary/truncated continuous) based on the latent Gaussian copula model.
}
\examples{

# Generate Data
set.seed(1)
n <- 100; p1 <- 5; p2 <- 7
maxcancor <- 0.9 # true canonical correlation
Sigma1 <- autocor(p1, 0.7)
blockind <- c(rep(1, 2), rep(2, p2-2))
Sigma2 <- blockcor(blockind, 0.7)
copula1 <- "exp"; copula2 <- "cube" # transformation for Gaussian copula model
mu <- rep(0, p1+p2)
type1 <- type2 <- "trunc" # X1: truncated continuous, X2: truncated continuous
c1 <- rep(0, p1); c2 <- rep(0, p2) # threshold for truncation of underlying continuous variable
trueidx1 <- c(0, 0, 1, 1, 1) # true variable indices for dataset X1
trueidx2 <- c(1, 0, 1, 0, 0, 1, 1) # true variable indices for dataset X2
simdata <- GenerateData(n=n, trueidx1 = trueidx1, trueidx2 = trueidx2, maxcancor = maxcancor,
                        Sigma1 = Sigma1, Sigma2 = Sigma2,
                        copula1 = copula1, copula2 = copula2,
                        muZ = mu,
                        type1 = type1, type2 = type2, c1 = c1, c2 = c2
)
X1 <- simdata$X1
X2 <- simdata$X2

# Estimate latent correlation matrix
estimateR(X1, type = "trunc")
estimateR_mixed(X1, X2, type1 = "trunc", type2 = "trunc")
}
\references{
Fan J., Liu H., Ning Y. and Zou H. (2017) \href{https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/rssb.12168}{"High dimensional semiparametric latent graphicalmodel for mixed data"}, \emph{J. R. Statist. Soc. B}, 79: 405-421.

Yoon G., Carroll R.J. and Gaynanova I. (2018+) \href{http://arxiv.org/abs/1807.05274}{"Sparse semiparametric canonical correlation analysis for data of mixed types"}, \emph{arXiv 1807.05274}.
}
