% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallCCA.R
\name{lambdaseq_generate}
\alias{lambdaseq_generate}
\title{Generate tuning parameter sequence for BIC.}
\usage{
lambdaseq_generate(nlambda = 100, initlam1 = NULL, initlam2 = NULL,
  lam.eps = 0.01, Sigma1, Sigma2, Sigma12, w1init = NULL, w2init = NULL)
}
\arguments{
\item{nlambda}{the number of tuning parameter sequence \code{lamseq} - default is 100.}

\item{initlam1}{initial value to generate a tuning parameter sequence for \eqn{X1}.}

\item{initlam2}{initial value to generate a tuning parameter sequence for \eqn{X2}.}

\item{lam.eps}{ratio of the smallest value for lambda to the maximum value of lambda.}

\item{Sigma1}{correlation matrix of \eqn{X1}.}

\item{Sigma2}{correlation matrix of \eqn{X2}.}

\item{Sigma12}{correlation matrix between \eqn{X1} and \eqn{X2}.}

\item{w1init}{initial value of canonical direction \eqn{w1}, which helps estimate \code{initlam1} better. The default is same coordinates for all elements such that \eqn{w^T \Sigma w = 1}.}

\item{w2init}{initial value of canonical direction \eqn{w2}, which helps estimate \code{initlam2} better. The default is same coordinates for all elements such that \eqn{w^T \Sigma w = 1}.}
}
\value{
a list containing two vectors of lambda sequences for \eqn{X1} and \eqn{X2}.
}
\description{
Generate tuning parameter sequence for BIC.
}
