% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallCCA.R
\name{myrcc}
\alias{myrcc}
\title{Canonical Ridge
This code is modified from rcc in the package \code{CCA}.}
\usage{
myrcc(X1, X2, lambda1, lambda2)
}
\arguments{
\item{X1}{numeric data matrix (n by p1).}

\item{X2}{numeric data matrix (n by p2).}

\item{lambda1}{regularization parameter sequence for \code{X1}. The length should be the same as \code{ncol(X1)}.}

\item{lambda2}{regularization parameter sequence for \code{X2}. The length should be the same as \code{ncol(X2)}.}
}
\value{
A data.frame containing
\itemize{
      \item{cancor: }{estimated canonical correlation.}
      \item{w1: }{estimated canonical direction \eqn{w1}.}
      \item{w2: }{estimated canonical direction \eqn{w2}.}
}
}
\description{
Canonical Ridge
This code is modified from rcc in the package \code{CCA}.
}
\seealso{
\code{\link{mixedCCA}}
}
