% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KendallCCA.R
\name{standardCCA}
\alias{standardCCA}
\title{Classical CCA using correlation matrix estimator.}
\usage{
standardCCA(S1, S2, S12, tol = 1e-04)
}
\arguments{
\item{S1}{A correlation matrix of \code{X1}.}

\item{S2}{A correlation matrix of \code{X2}.}

\item{S12}{A correlation matrix between \code{X1} and \code{X2}.}

\item{tol}{A cutoff for eigenvalues. During the calculation, eigenvalues smaller than cutoff are set to zero.}
}
\value{
\code{standardCCA} returns a data.frame containing
\itemize{
      \item{cancor: }{estimated canonical correlation.}
      \item{w1: }{estimated canonical direction \eqn{w1}.}
      \item{w2: }{estimated canonical direction \eqn{w2}.}
}
}
\description{
Conduct the canonical correlation analysis (CCA) between two datasets using their correlation matrices as an input. Only first canonical pair with the largest canonical correlation is calculated.
}
\seealso{
See \code{\link{mixedCCA}} for examples.
}
